refactor: extract fetch
